// src/scrapers/interfaces/find-one-response.ts
import { ParsingModel } from "./find-many-response"

export interface FindOneScraperResponse {
    id: string
    createdAt: string
    updatedAt: string
    name: string
    defaultUrl: string | null
    description: string | null
    ownerId: string
    clientId: string
    parsingModel: ParsingModel
}